sepeare Jelly into DiffuseMissile and Jelly
